requirements.txt
runpod>=1.7.0
requests>=2.31.0
Pillow>=10.0.0
